Replace Writer with State to avoid repeating work
This is a pretty ugly first pass, it boots performance significantly, so
in it goes until I can make it nicer to read.

It is also a little specific to large lists of repeating elements. The idea
is to store an IntMap of rendered classes, if we've rendered that
specific set of class names before we should try not to do it again, as
it appears to be expensive.

I might remove the use of DList after this change as it might make
hashing the styles to be rendered as classes a little easier ... not sure.

Anyway, this gets MisoStyle into a 1.03 : 1.47 relationship with
un-adorned Miso in the JS Benchmark. I'm happy enough with this :).